refactor: This enhancement includes modularization of the exception elements modules, in order to apply more cohesion and to separate the local from the base
