+++ devdox/app/exceptions/base_exceptions.py
@@ -10,7 +10,7 @@ Key responsibilities:
 - Centralized support for structured API responses and log formatting
 
 To create a new custom exception:
-1. Subclass `DevDoxAPIException`
+1. Subclass `DevDoxAPIException` in a different module preferably local to your project
 2. Optionally override `http_status`
 3. Pass `user_message`, `log_message`, `error_type`, etc.
 
@@ -25,29 +25,13 @@ Example:
                 internal_context={"repo_id": repo_id},
             )
 
-These exceptions are globally handled and logged by the `devdox_exception_handler`
-defined in `handlers.py`, and registered into the FastAPI app via `register.py`.
+These exceptions are globally handled and logged usually by a handler.
 """
-import dataclasses
+
 from typing import Any, Dict, Optional
 
 from starlette import status
 
-from app.exceptions.exception_constants import (
-    AUTH_FAILED,
-    GENERIC_BAD_REQUEST,
-    GENERIC_RESOURCE_NOT_FOUND,
-    GENERIC_VALIDATION_FAILED_USER_MESSAGE,
-)
-
-
-@dataclasses.dataclass
-class ErrorPayload:
-    message: str
-    status_code: int
-    details: Optional[Dict[str, Any]] = None
-    debug: Optional[Any] = None
-    error_type: Optional[str] = None
 
 class DevDoxAPIException(Exception):
     """
@@ -98,6 +82,7 @@ class DevDoxAPIException(Exception):
 
                 error_type: Optional machine-readable code.
                 http_status_override: Override the default HTTP status for this exception.
+                log_level: specifies the level of the logging system of exception instance
         """
         super().__init__(user_message)
 
@@ -111,46 +96,3 @@ class DevDoxAPIException(Exception):
 
     def __str__(self):
         return f"[{self.error_type}] {self.user_message}"
-
-class UnauthorizedAccess(DevDoxAPIException):
-    http_status = status.HTTP_401_UNAUTHORIZED
-
-    def __init__(self, reason=None, log_message=None, log_level=None):
-        
-        if not reason or not reason.strip():
-            reason = AUTH_FAILED
-        
-        super().__init__(
-            user_message=reason, log_message=log_message, log_level=log_level
-        )
-
-
-class BadRequest(DevDoxAPIException):
-    http_status = status.HTTP_400_BAD_REQUEST
-
-    def __init__(self, reason=None, log_message: Optional[str] = None):
-        
-        if not reason or not reason.strip():
-            reason = GENERIC_BAD_REQUEST
-        
-        super().__init__(user_message=reason, log_message=log_message, log_level="warning")
-
-
-class ResourceNotFound(DevDoxAPIException):
-    http_status = status.HTTP_404_NOT_FOUND
-
-    def __init__(self, reason=None):
-        
-        if not reason or not reason.strip():
-            reason = GENERIC_RESOURCE_NOT_FOUND
-        
-        super().__init__(user_message=reason)
-
-class ValidationFailed(DevDoxAPIException):
-    http_status = status.HTTP_400_BAD_REQUEST
-
-    def __init__(self, field_errors: Dict[str, list]):
-        super().__init__(
-            user_message=GENERIC_VALIDATION_FAILED_USER_MESSAGE,
-            public_context=field_errors
-        )
