+++ devdox/app/services/git_tokens.py
@@ -5,7 +5,7 @@ from typing import Annotated, Optional
 from fastapi import Depends
 from tortoise.exceptions import IntegrityError
 
-from app.exceptions.custom_exceptions import BadRequest, ResourceNotFound
+from app.exceptions.local_exceptions import BadRequest, ResourceNotFound
 from app.exceptions.exception_constants import (
     GENERIC_ALREADY_EXIST,
     TOKEN_MISSING,
