+++ devdox/app/services/repository.py
@@ -3,7 +3,7 @@ from uuid import UUID, uuid4
 from fastapi import Depends
 from tortoise.exceptions import IntegrityError
 
-from app.exceptions.custom_exceptions import (
+from app.exceptions.local_exceptions import (
     BadRequest,
     ResourceNotFound,
 )
