+++ devdox/tests/unit_test/app/services/test_api_keys_service.py
@@ -7,7 +7,7 @@ from uuid import uuid4
 import pytest
 from pydantic import ValidationError
 
-from app.exceptions.custom_exceptions import BadRequest, ResourceNotFound
+from app.exceptions.local_exceptions import BadRequest, ResourceNotFound
 from app.exceptions.exception_constants import (
     FAILED_GENERATE_API_KEY_RETRIES_LOG_MESSAGE,
     UNIQUE_API_KEY_GENERATION_FAILED,
